requirements.txt
aiofiles==0.8.0
fastapi==0.78.0
python_crontab==2.6.0
uvicorn[standard]
